+++ dotenv/mod.ts
@@ -176,9 +176,10 @@
 
 import { parse } from "./parse.ts";
 
-export * from "./prepare.ts";
 export * from "./stringify.ts";
 export * from "./parse.ts";
+export * from "./assert_env_set.ts";
+export * from "./set_env_default.ts";
 
 export interface LoadOptions {
   /**
